# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HDP App OpenApi API Documentation",
        "description": "HDP App Using L5 Swagger OpenApi description",
        "contact": {
            "email": "hdp@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/facility/facility-user": {
            "post": {
                "tags": [
                    "facility"
                ],
                "summary": "create a facility product offering ",
                "description": "create a facility product offering ",
                "operationId": "createFacilityUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FacilityUserCreateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/facility/facility-user/{id}": {
            "get": {
                "tags": [
                    "facility"
                ],
                "summary": "show a facility's dependant ",
                "description": "show a facility's dependant ",
                "operationId": "showFacilityUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Dependant ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "facility"
                ],
                "summary": "update a facility's dependant ",
                "description": "update a facility's dependant ",
                "operationId": "updateFacilityUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Dependant ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "facility"
                ],
                "summary": "delete a facility's dependant ",
                "description": "delete a facility's dependant ",
                "operationId": "deleteFacilityUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Dependant ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/general/general-service-offerings": {
            "get": {
                "tags": [
                    "general"
                ],
                "summary": "Show all service-offerings ",
                "description": "Show all service-offerings ",
                "operationId": "allServiceOfferings",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/general/general-service-offerings/{id}": {
            "get": {
                "tags": [
                    "general"
                ],
                "summary": "Show facility details ",
                "description": "Show facility details ",
                "operationId": "showServiceOfferingDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Facility ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/professional/{id}/review/": {
            "get": {
                "tags": [
                    "professional"
                ],
                "summary": "show a professional's review details ",
                "description": "show a professional's review details ",
                "operationId": "showProfessionalReview",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Review ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FacilityUserCreateFormRequest": {
                "title": "Facility User Create Form Request Fields",
                "description": "Facility User Create request body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "facility_id": {
                        "title": "Facility Id",
                        "description": "Facility Id",
                        "type": "integer",
                        "example": "1"
                    },
                    "offer_name": {
                        "title": " Facility Public code",
                        "description": " Facility Public code",
                        "type": "string",
                        "example": "name"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}